+++ src/main/java/com/iexec/core/replicate/ReplicatesService.java
@@ -19,6 +19,7 @@ package com.iexec.core.replicate;
 import com.iexec.common.replicate.*;
 import com.iexec.common.utils.ContextualLockRunner;
 import com.iexec.commons.poco.chain.ChainContribution;
+import com.iexec.commons.poco.chain.ChainTask;
 import com.iexec.commons.poco.notification.TaskNotificationType;
 import com.iexec.commons.poco.task.TaskDescription;
 import com.iexec.core.chain.IexecHubService;
@@ -30,10 +31,15 @@ import io.vavr.control.Either;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.dao.OptimisticLockingFailureException;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.retry.annotation.Recover;
 import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
+import org.web3j.utils.Numeric;
 
 import java.util.Collections;
 import java.util.List;
@@ -47,6 +53,7 @@ import static com.iexec.common.replicate.ReplicateStatusCause.REVEAL_TIMEOUT;
 @Service
 public class ReplicatesService {
 
+    private final MongoTemplate mongoTemplate;
     private final ReplicatesRepository replicatesRepository;
     private final IexecHubService iexecHubService;
     private final ApplicationEventPublisher applicationEventPublisher;
@@ -57,12 +64,14 @@ public class ReplicatesService {
     private final ContextualLockRunner<String> replicatesUpdateLockRunner =
             new ContextualLockRunner<>(10, TimeUnit.MINUTES);
 
-    public ReplicatesService(ReplicatesRepository replicatesRepository,
+    public ReplicatesService(MongoTemplate mongoTemplate,
+                             ReplicatesRepository replicatesRepository,
                              IexecHubService iexecHubService,
                              ApplicationEventPublisher applicationEventPublisher,
                              Web3jService web3jService,
                              ResultService resultService,
                              TaskLogsService taskLogsService) {
+        this.mongoTemplate = mongoTemplate;
         this.replicatesRepository = replicatesRepository;
         this.iexecHubService = iexecHubService;
         this.applicationEventPublisher = applicationEventPublisher;
@@ -101,26 +110,16 @@ public class ReplicatesService {
     }
 
     public List<Replicate> getReplicates(String chainTaskId) {
-        Optional<ReplicatesList> optionalList = getReplicatesList(chainTaskId);
-        if (optionalList.isEmpty()) {
-            return Collections.emptyList();
-        }
-        return optionalList.get().getReplicates();
+        return getReplicatesList(chainTaskId)
+                .map(ReplicatesList::getReplicates)
+                .orElse(Collections.emptyList());
     }
 
     public Optional<Replicate> getReplicate(String chainTaskId, String walletAddress) {
-        Optional<ReplicatesList> optional = getReplicatesList(chainTaskId);
-        if (optional.isEmpty()) {
-            return Optional.empty();
-        }
-
-        for (Replicate replicate : optional.get().getReplicates()) {
-            if (replicate.getWalletAddress().equals(walletAddress)) {
-                return Optional.of(replicate);
-            }
-        }
-
-        return Optional.empty();
+        return getReplicates(chainTaskId)
+                .stream()
+                .filter(replicate -> replicate.getWalletAddress().equals(walletAddress))
+                .findFirst();
     }
 
     public int getNbReplicatesWithCurrentStatus(String chainTaskId, ReplicateStatus... listStatus) {
@@ -248,6 +247,15 @@ public class ReplicatesService {
                 resultLink = details.getResultLink();
                 chainCallbackData = details.getChainCallbackData();
             }
+        }
+
+        if (statusUpdate.getStatus() == CONTRIBUTE_AND_FINALIZE_DONE) {
+            // TODO read chainCallbackData if CONTRIBUTE_AND_FINALIZE becomes applicable some day in the future and if latest ABI is used
+            resultLink = iexecHubService.getChainTask(chainTaskId)
+                    .map(ChainTask::getResults)
+                    .map(Numeric::hexStringToByteArray)
+                    .map(String::new)
+                    .orElse(null);
         }
 
         return UpdateReplicateStatusArgs.builder()
@@ -402,16 +410,14 @@ public class ReplicatesService {
         log.info("Replicate update request [status:{}, chainTaskId:{}, walletAddress:{}, details:{}]",
                 statusUpdate.getStatus(), chainTaskId, walletAddress, statusUpdate.getDetailsWithoutLogs());
 
-        final Optional<ReplicatesList> oReplicatesList = getReplicatesList(chainTaskId);
-        final Optional<Replicate> oReplicate = oReplicatesList
-                .flatMap(replicatesList -> replicatesList.getReplicateOfWorker(walletAddress));
-        if (oReplicatesList.isEmpty() || oReplicate.isEmpty()) {
+        final Replicate replicate = getReplicatesList(chainTaskId)
+                .flatMap(replicatesList -> replicatesList.getReplicateOfWorker(walletAddress))
+                .orElse(null);
+        if (replicate == null) {
             log.error("Cannot update replicate, could not get replicate [chainTaskId:{}, UpdateRequest:{}]",
                     chainTaskId, statusUpdate);
             return Either.left(ReplicateStatusUpdateError.UNKNOWN_REPLICATE);
         }
-        final ReplicatesList replicatesList = oReplicatesList.get();
-        final Replicate replicate = oReplicate.get();
         final ReplicateStatus newStatus = statusUpdate.getStatus();
 
         final ReplicateStatusUpdateError error = canUpdateReplicateStatus(replicate, statusUpdate, updateReplicateStatusArgs);
@@ -419,14 +425,19 @@ public class ReplicatesService {
             return Either.left(error);
         }
 
+        Update update = new Update();
         if (newStatus == CONTRIBUTED || newStatus == CONTRIBUTE_AND_FINALIZE_DONE) {
             replicate.setContributionHash(updateReplicateStatusArgs.getChainContribution().getResultHash());
             replicate.setWorkerWeight(updateReplicateStatusArgs.getWorkerWeight());
+            update.set("replicates.$.contributionHash", updateReplicateStatusArgs.getChainContribution().getResultHash());
+            update.set("replicates.$.workerWeight", updateReplicateStatusArgs.getWorkerWeight());
         }
 
         if (newStatus == RESULT_UPLOADED || newStatus == CONTRIBUTE_AND_FINALIZE_DONE) {
             replicate.setResultLink(updateReplicateStatusArgs.getResultLink());
             replicate.setChainCallbackData(updateReplicateStatusArgs.getChainCallbackData());
+            update.set("replicates.$.resultLink", updateReplicateStatusArgs.getResultLink());
+            update.set("replicates.$.chainCallbackData", updateReplicateStatusArgs.getChainCallbackData());
         }
 
         if (statusUpdate.getDetails() != null &&
@@ -436,10 +447,13 @@ public class ReplicatesService {
             taskLogsService.addComputeLogs(chainTaskId, computeLogs);
             statusUpdate.getDetails().setComputeLogs(null);//using null here to keep light replicate
             replicate.setAppComputeLogsPresent(true);
+            update.set("replicates.$.appComputeLogsPresent", true);
         }
 
+        update.push("replicates.$.statusUpdateList", statusUpdate);
+        Query query = Query.query(Criteria.where("chainTaskId").is(chainTaskId).and("replicates.walletAddress").is(walletAddress));
+        mongoTemplate.updateFirst(query, update, ReplicatesList.class);
         replicate.updateStatus(statusUpdate);
-        replicatesRepository.save(replicatesList);
         applicationEventPublisher.publishEvent(new ReplicateUpdatedEvent(chainTaskId, walletAddress, statusUpdate));
         ReplicateStatusCause newStatusCause = statusUpdate.getDetails() != null ?
                 statusUpdate.getDetails().getCause() : null;